backend/requirements.txt
fastapi
uvicorn
openai
requests
python-dotenv
pymongo
motor
openai
google-generativeai
jinja2


